Fix local CI theme lint failure
