+++ web/src/pages/testUtils.tsx
@@ -2,7 +2,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { render } from '@testing-library/react';
 import { RouterProvider, createMemoryRouter } from 'react-router-dom';
 import { vi } from 'vitest';
-import { ThemeProvider } from '../hooks/useTheme';
+import { ThemeProvider } from '../hooks/ThemeProvider';
 import {
   OTHER_SESSION_EXTERNAL_ID,
   OTHER_SESSION_ID,
